Update device.mk, BoardConfig.mk, and device_framework_matrix.xml
This commit updates the device.mk file to remove the IMS inheritance and adds the necessary IMS boot jars. It also modifies the BoardConfig.mk file to include the vendor and odm partitions and changes the file system type for vendor and odm to erofs. Additionally, a new device_framework_matrix.xml file is added.